+++ tergite_autocalibration/lib/base/utils/figure_utils.py
@@ -23,9 +23,7 @@ from tergite_autocalibration.config.globals import (
 from tergite_autocalibration.utils.logging import logger
 
 
-def _add_top_band(
-    axes_tuple,
-):
+def _add_top_band(axes_tuple):
 
     ax_left, ax_center, ax_right = axes_tuple
 
@@ -48,7 +46,7 @@ def _add_top_band(
     ax_center.add_line(line)
 
     # Add left logo, this is hardcoded as it is the Tergite logo.
-    logo_path = ("resources/logo.png",)
+    logo_path = "resources/logo.png"
 
     if logo_path:
         try:
@@ -187,8 +185,6 @@ def create_figure_with_top_band(nrows, ncols) -> tuple:
 
     top_band_axes = (ax_left, ax_center, ax_right)
 
-    _add_top_band(
-        top_band_axes,
-    )
+    _add_top_band(top_band_axes)
 
     return fig, axs