+++ tergite_autocalibration/lib/nodes/readout/punchout/node.py
@@ -23,6 +23,11 @@ from tergite_autocalibration.lib.utils.samplespace import resonator_samples
 
 
 class PunchoutNode(ScheduleQubitNode):
+    """
+    This class implements the punchout node, which is used to measure the
+    readout amplitude.
+    """
+
     measurement_obj = PunchoutMeasurement
     analysis_obj = PunchoutNodeAnalysis
     qubit_qois = ["measure:pulse_amp"]
@@ -35,6 +40,6 @@ class PunchoutNode(ScheduleQubitNode):
                 qubit: resonator_samples(qubit) for qubit in self.all_qubits
             },
             "ro_amplitudes": {
-                qubit: np.linspace(0.008, 0.04, 12) for qubit in self.all_qubits
+                qubit: np.linspace(0.004, 0.04, 20) for qubit in self.all_qubits
             },
         }